Refactor(Startup): Optimize app initialization and merge entry points
This commit refactors the application startup flow to reduce async dependency chains, consolidates the main UI entry point, and removes the intermediate loading screen.

- **refactor(Repositories):** Converted `AppDataRepository` and `UserManagement` instantiation to synchronous factory methods by injecting `SharedPreferences` from `main.dart`. Introduced `initialize()` methods for subsequent async setup.
- **refactor(MasterPage):** Merged `WandrrApp` logic into `MasterPage` and deleted `wandrr_app.dart`. Removed the Rive "Walk" animation loading screen to allow for immediate UI rendering.
- **refactor(Bloc):** Updated `MasterPageBloc` to initialize with a `LoadedRepository` state immediately and trigger async initialization via `_StartupInternal`.
- **refactor(Auth):** Removed redundant `_isLoggedIn` flag usage from `UserManagement` persistence logic.
- **chore(main):** Moved `Firebase.initializeApp` and `SharedPreferences.getInstance` to `main.dart` to support synchronous dependency injection.
